web_app/requirements.txt
Flask==3.1.0
numpy==1.24.3
Pillow==11.2.1
tensorflow==2.17.0
tensorflow_intel==2.17.0
